Структурированы файлы проекта по фичам
