--- TestTask/CharType.cs
@@ -1,18 +1,0 @@
-﻿namespace TestTask
-{
-    /// <summary>
-    /// Тип букв
-    /// </summary>
-    public enum CharType
-    {
-        /// <summary>
-        /// Гласные
-        /// </summary>
-        Vowel,
-
-        /// <summary>
-        /// Согласные
-        /// </summary>
-        Consonants
-    }
-}
