+++ TestTask/LetterAnalysis/CharType.cs
@@ -1,0 +1,18 @@
+﻿namespace TestTask
+{
+    /// <summary>
+    /// Тип букв
+    /// </summary>
+    public enum CharType
+    {
+        /// <summary>
+        /// Гласные
+        /// </summary>
+        Vowel,
+
+        /// <summary>
+        /// Согласные
+        /// </summary>
+        Consonants
+    }
+}
