+++ TestTask/LetterAnalysis/DictionaryLetterStorage.cs
