+++ TestTask/LetterAnalysis/ILetterAnalysisStorage.cs
