+++ TestTask/LetterAnalysis/LetterStats.cs
@@ -1,0 +1,18 @@
+﻿namespace TestTask
+{
+    /// <summary>
+    /// Статистика вхождения буквы/пары букв
+    /// </summary>
+    public struct LetterStats
+    {
+        /// <summary>
+        /// Буква/Пара букв для учёта статистики.
+        /// </summary>
+        public string Letter;
+
+        /// <summary>
+        /// Кол-во вхождений буквы/пары.
+        /// </summary>
+        public int Count;
+    }
+}
