+++ TestTask/ReadOnlyStream/IReadOnlyStreamFactory.cs
