+++ TestTask/Utilities/ConsoleHelperMethods.cs
